# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DOCX Exporter REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/document-language": {
      "get": {
        "operationId": "getDocumentLanguage",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "spaceId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "documentName",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved document language"
          }
        },
        "summary": "Gets language of specified Polarion's document, defined in its custom field 'docLanguage'",
        "tags": [
          "Utility resources"
        ]
      }
    },
    "/api/export-filename": {
      "post": {
        "operationId": "getFileName",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParams"
              }
            }
          },
          "description": "Export parameters similar as for generation of PDF",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully generated filename"
          }
        },
        "summary": "Gets a filename, prepared with velocity and placeholders",
        "tags": [
          "Utility resources"
        ]
      }
    },
    "/api/link-role-names": {
      "get": {
        "operationId": "readLinkRoleNames",
        "parameters": [
          {
            "description": "Project scope in form project/<PROJECT_ID>/",
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully retrieved list of link role names"
          }
        },
        "summary": "Gets list of possible WorkItem link role names in specified project scope",
        "tags": [
          "Utility resources"
        ]
      }
    },
...
    "/api/projects/{projectId}/name": {
      "get": {
        "operationId": "getProjectName",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved project name"
          }
        },
        "summary": "Gets name of specified Polarion's project",
        "tags": [
          "Utility resources"
        ]
      }
    },
    "/api/projects/{projectId}/testruns/{testRunId}/attachments": {
      "get": {
        "operationId": "getTestRunAttachments",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filename filter for attachment",
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TestRunAttachment"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully retrieved list of attachments"
          }
        },
        "summary": "Gets list of attachments of specified workitem using optional filter",
        "tags": [
          "TestRuns"
        ]
      }
    },
    "/api/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}": {
      "get": {
        "operationId": "getTesRunAttachment",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment ID",
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunAttachment"
                }
              }
            },
            "description": "Successfully retrieved attachment"
          }
        },
        "summary": "Gets attachment of specified workitem",
        "tags": [
          "TestRuns"
        ]
      }
    },
    "/api/projects/{projectId}/testruns/{testRunId}/attachments/{attachmentId}/content": {
      "get": {
        "operationId": "getTestRunAttachmentContent",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "TestRun ID",
            "in": "path",
            "name": "testRunId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment ID",
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Attachment revision",
            "in": "query",
            "name": "revision",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved attachment content"
          }
        },
        "summary": "Gets content of attachment of specified workitem",
        "tags": [
          "TestRuns"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportParams": {
        "properties": {
          "baselineRevision": {
            "description": "The specific revision of the provided baseline",
            "type": "string"
          },
          "chapters": {
            "description": "Specific higher level chapters",
            "items": {
              "description": "Specific higher level chapters",
              "type": "string"
            },
            "type": "array"
          },
          "cutEmptyChapters": {
            "description": "Empty chapters should be removed from the document",
            "type": "boolean"
          },
          "cutEmptyWIAttributes": {
            "default": true,
            "description": "Empty work item attributes should be removed from the document",
            "type": "boolean"
          },
          "cutLocalUrls": {
            "description": "Local Polarion URLs should be removed from the document",
            "type": "boolean"
          },
          "fileName": {
            "description": "Target file name of exported document",
            "type": "string"
          },
          "includeUnreferencedComments": {
            "description": "Include unreferenced comments",
            "type": "boolean"
          },
          "internalContent": {
            "description": "Internal content",
            "type": "string"
          },
          "language": {
            "description": "Language in the exported document",
            "type": "string"
          },
          "linkRoleDirection": {
            "default": "BOTH",
            "description": "Direction of selected link roles to include",
            "enum": [
              "BOTH",
              "DIRECT",
              "REVERSE"
            ],
            "type": "string"
          },
          "linkedWorkitemRoles": {
            "description": "Specific Workitem roles",
            "example": [
              "has parent",
              "depends on"
            ],
            "items": {
              "description": "Specific Workitem roles",
              "example": "[\"has parent\",\"depends on\"]",
              "type": "string"
            },
            "type": "array"
          },
          "localization": {
            "description": "Localization settings name",
            "type": "string"
          },
          "locationPath": {
            "description": "Document path for export",
            "example": "Specification/Product Specification",
            "type": "string"
          },
          "orientation": {
            "description": "Page orientation",
            "type": "string"
          },
          "paperSize": {
            "description": "Paper size",
            "type": "string"
          },
          "projectId": {
            "description": "The unique identifier for the project",
            "example": "elibrary",
            "type": "string"
          },
          "removalSelector": {
            "description": "CSS-like selector for elements to remove",
            "type": "string"
          },
          "renderComments": {
            "description": "Which comments should be rendered in the exported document",
            "enum": [
              "OPEN",
              "ALL"
            ],
            "type": "string"
          },
          "revision": {
            "description": "The specific revision of the document",
            "type": "string"
          },
          "template": {
            "description": "Template settings name",
            "type": "string"
          },
          "urlQueryParameters": {
            "additionalProperties": {
              "description": "Map of attributes extracted from the URL",
              "type": "string"
            },
            "description": "Map of attributes extracted from the URL",
            "type": "object"
          },
          "webhooks": {
            "description": "Webhooks settings name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TestRunAttachment": {
        "properties": {
          "fileName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
...
}