# src/main/java/ch/sbb/polarion/extension/docx_exporter/rest/controller/UtilityResourcesInternalController.java
package ch.sbb.polarion.extension.docx_exporter.rest.controller;

import ch.sbb.polarion.extension.docx_exporter.model.WebhooksStatus;
import ch.sbb.polarion.extension.docx_exporter.properties.DocxExporterExtensionConfiguration;
import ch.sbb.polarion.extension.docx_exporter.rest.model.conversion.ExportParams;
import ch.sbb.polarion.extension.docx_exporter.service.DocxExporterPolarionService;
import ch.sbb.polarion.extension.docx_exporter.util.DocumentFileNameHelper;
import ch.sbb.polarion.extension.docx_exporter.util.EnumValuesProvider;
import com.polarion.alm.tracker.model.IModule;
import com.polarion.alm.tracker.model.ITrackerProject;
import com.polarion.platform.persistence.IEnumOption;
import io.swagger.v3.oas.annotations.Hidden;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.parameters.RequestBody;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;

import jakarta.inject.Singleton;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import java.util.Collections;
import java.util.List;

import static ch.sbb.polarion.extension.docx_exporter.util.placeholder.PlaceholderValues.DOC_LANGUAGE_FIELD;

@Singleton
@Hidden
@Path("/internal")
@Tag(name = "Utility resources")
public class UtilityResourcesInternalController {

    private final DocxExporterPolarionService docxExporterPolarionService;

    public UtilityResourcesInternalController() {
        docxExporterPolarionService = new DocxExporterPolarionService();
    }

    public UtilityResourcesInternalController(DocxExporterPolarionService docxExporterPolarionService) {
        this.docxExporterPolarionService = docxExporterPolarionService;
    }

    @GET
    @Path("/link-role-names")
    @Produces(MediaType.APPLICATION_JSON)
    @Operation(summary = "Gets list of possible WorkItem link role names in specified project scope",
            responses = {
                    @ApiResponse(responseCode = "200",
                            description = "Successfully retrieved list of link role names",
                            content = @Content(array = @ArraySchema(schema = @Schema(implementation = String.class)))
                    )
            }
    )
    public List<String> readLinkRoleNames(@Parameter(description = "Project scope in form project/<PROJECT_ID>/", required = true) @QueryParam("scope") String scope) {
        ITrackerProject project = docxExporterPolarionService.getProjectFromScope(scope);
        if (project != null) {
            return EnumValuesProvider.getAllLinkRoleNames(project);
        }
        return Collections.emptyList();
    }

    @GET
    @Path("/document-language")
    @Produces(MediaType.TEXT_PLAIN)
    @Operation(summary = "Gets language of specified Polarion's document, defined in its custom field 'docLanguage'",
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successfully retrieved document language")
            }
    )
    public String getDocumentLanguage(@QueryParam("projectId") String projectId, @QueryParam("spaceId") String spaceId,
                                      @QueryParam("documentName") String documentName, @QueryParam("revision") String revision) {
        IModule module = docxExporterPolarionService.getModule(projectId, spaceId, documentName, revision);
        Object documentLanguageField = module.getCustomField(DOC_LANGUAGE_FIELD);
        return (documentLanguageField instanceof IEnumOption option) ? option.getId() : null;
    }

    @GET
    @Path("/projects/{projectId}/name")
    @Produces(MediaType.TEXT_PLAIN)
    @Operation(summary = "Gets name of specified Polarion's project",
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successfully retrieved project name")
            }
    )
    public String getProjectName(@PathParam("projectId") String projectId) {
        return docxExporterPolarionService.getProject(projectId).getName();
    }

    @POST
    @Path("/export-filename")
    @Consumes(MediaType.APPLICATION_JSON)
    @Produces(MediaType.TEXT_PLAIN)
    @Operation(summary = "Gets a filename, prepared with velocity and placeholders",
            requestBody = @RequestBody(description = "Export parameters similar as for generation of PDF",
                    required = true,
                    content = @Content(schema = @Schema(implementation = ExportParams.class),
                            mediaType = MediaType.APPLICATION_JSON
                    )
            ),
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successfully generated filename")
            }
    )
    public String getFileName(ExportParams exportParams) {
        return new DocumentFileNameHelper().getDocumentFileName(exportParams);
    }

    @GET
    @Path("/webhooks/status")
    @Produces(MediaType.APPLICATION_JSON)
    @Tag(name = "Utility resources")
    @Operation(summary = "Gets webhooks status - if they are enabled or not",
            responses = {
                    @ApiResponse(responseCode = "200", description = "Webhooks status",
                            content = @Content(schema = @Schema(implementation = WebhooksStatus.class)))
            }
    )
    public WebhooksStatus getWebhooksStatus() {
        Boolean webhooksEnabled = DocxExporterExtensionConfiguration.getInstance().getWebhooksEnabled();
        return WebhooksStatus.builder().enabled(webhooksEnabled).build();
    }
}